+++ beacon-chain/rpc/apimiddleware/custom_hooks_test.go
@@ -360,31 +360,6 @@ func TestPreparePublishedBlindedBlock(t *testing.T) {
 	})
 }
 
-func TestPrepareValidatorAggregates(t *testing.T) {
-	body := &tempSyncCommitteesResponseJson{
-		Data: &tempSyncCommitteeValidatorsJson{
-			Validators: []string{"1", "2"},
-			ValidatorAggregates: []*tempSyncSubcommitteeValidatorsJson{
-				{
-					Validators: []string{"3", "4"},
-				},
-				{
-					Validators: []string{"5"},
-				},
-			},
-		},
-	}
-	bodyJson, err := json.Marshal(body)
-	require.NoError(t, err)
-
-	container := &SyncCommitteesResponseJson{}
-	runDefault, errJson := prepareValidatorAggregates(bodyJson, container)
-	require.Equal(t, nil, errJson)
-	require.Equal(t, apimiddleware.RunDefault(false), runDefault)
-	assert.DeepEqual(t, []string{"1", "2"}, container.Data.Validators)
-	require.DeepEqual(t, [][]string{{"3", "4"}, {"5"}}, container.Data.ValidatorAggregates)
-}
-
 func TestSerializeV2Block(t *testing.T) {
 	t.Run("Phase 0", func(t *testing.T) {
 		response := &BlockV2ResponseJson{
