+++ beacon-chain/rpc/eth/beacon/BUILD.bazel
@@ -8,13 +8,12 @@ go_library(
         "config.go",
         "handlers.go",
         "handlers_pool.go",
+        "handlers_state.go",
         "handlers_validator.go",
         "log.go",
         "pool.go",
         "server.go",
-        "state.go",
         "structs.go",
-        "sync_committee.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/beacon",
     visibility = ["//visibility:public"],
@@ -83,13 +82,12 @@ go_test(
         "blocks_test.go",
         "config_test.go",
         "handlers_pool_test.go",
+        "handlers_state_test.go",
         "handlers_test.go",
         "handlers_validators_test.go",
         "init_test.go",
         "pool_test.go",
         "server_test.go",
-        "state_test.go",
-        "sync_committee_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
@@ -149,8 +147,6 @@ go_test(
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@com_github_stretchr_testify//mock:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//codes:go_default_library",
-        "@org_golang_google_grpc//status:go_default_library",
         "@org_golang_google_protobuf//types/known/emptypb:go_default_library",
     ],
 )
